Move version views to seperate file
