+++ python/aristotle-metadata-registry/aristotle_mdr/views/views.py
@@ -15,23 +15,20 @@ from django.utils.translation import ugettext_lazy as _
 from django.views.generic import TemplateView, RedirectView
 from django.utils.decorators import method_decorator
 from django.utils.module_loading import import_string
-from django.utils.dateparse import parse_datetime
 from django.contrib.contenttypes.models import ContentType
 from formtools.wizard.views import SessionWizardView
 
 import json
 import copy
-from collections import defaultdict
 
 import reversion
-from reversion_compare.views import HistoryCompareDetailView
 
 from aristotle_mdr.perms import (
     user_can_view, user_can_edit,
     user_can_change_status
 )
 from aristotle_mdr import perms
-from aristotle_mdr.utils import url_slugify_concept, pretify_camel_case
+from aristotle_mdr.utils import url_slugify_concept
 
 from aristotle_mdr import forms as MDRForms
 from aristotle_mdr import models as MDR
@@ -39,11 +36,9 @@ from aristotle_mdr.utils import get_concepts_for_apps, fetch_aristotle_settings,
 from aristotle_mdr.views.utils import generate_visibility_matrix, CachePerItemUserMixin
 from aristotle_mdr.contrib.slots.utils import get_allowed_slots
 from aristotle_mdr.contrib.favourites.models import Favourite, Tag
-from aristotle_mdr.managers import ConceptManager
 
 from haystack.views import FacetedSearchView
 
-from ckeditor_uploader.fields import RichTextUploadingField as RichTextField
 
 import logging
 
@@ -68,50 +63,6 @@ class SmartRoot(RedirectView):
 class DynamicTemplateView(TemplateView):
     def get_template_names(self):
         return ['aristotle_mdr/static/%s.html' % self.kwargs['template']]
-
-
-class ConceptHistoryCompareView(HistoryCompareDetailView):
-    model = MDR._concept
-    pk_url_kwarg = 'iid'
-    template_name = "aristotle_mdr/actions/concept_history_compare.html"
-
-    compare_exclude = [
-        'favourites',
-        'user_view_history',
-        'submitter',
-        'is_public',
-        'is_locked'
-    ]
-
-    item_action_url = 'aristotle:item_version'
-
-    def get_object(self, queryset=None):
-        item = super().get_object(queryset)
-        if not user_can_view(self.request.user, item):
-            raise PermissionDenied
-        self.model = item.item.__class__  # Get the subclassed object
-        return item
-
-    # Overwrite this to add item url
-    def _get_action_list(self):
-        action_list = []
-        versions = self._order_version_queryset(
-            reversion.models.Version.objects.get_for_object(
-                self.get_object()
-            ).select_related("revision__user")
-        )
-
-        for version in versions:
-            action_list.append({
-                'version': version,
-                'revision': version.revision,
-                'url': reverse(self.item_action_url, args=[version.id])
-            })
-        return action_list
-
-    @method_decorator(login_required)
-    def dispatch(self, *args, **kwargs):
-        return super().dispatch(*args, **kwargs)
 
 
 def notification_redirect(self, content_type, object_id):
@@ -349,388 +300,6 @@ class DataElementView(ConceptRenderMixin, TemplateView):
             'statuses'
         ]
         return model.objects.select_related(*related_objects).prefetch_related(*prefetch_objects)
-
-
-class VersionField:
-
-    perm_message = 'Linked to object(s) you do not have permission to view'
-
-    def __init__(self, value='', obj=None, help_text='', html=False, reference_label=''):
-        if not value:
-            self.value = ''
-        else:
-            self.value=str(value)
-
-        if type(obj) == list and len(obj) == 1:
-            self.obj = obj[0]
-        else:
-            self.obj=obj
-
-        self.help_text=help_text
-        self.is_html=html
-        self.reference_label = reference_label
-
-    def dereference(self, lookup):
-        if self.is_reference and self.reference_label in lookup:
-            id_lookup = lookup[self.reference_label]
-            if self.is_list:
-                deref_list = []
-                for pk in self.obj:
-                    if pk in id_lookup:
-                        deref_list.append(id_lookup[pk])
-                self.obj = deref_list
-            else:
-                if self.obj in id_lookup:
-                    self.obj = id_lookup[self.obj]
-                else:
-                    self.obj = None
-                    self.value = self.perm_message
-
-        self.reference_label = ''
-
-    @property
-    def is_reference(self):
-        return self.reference_label is not ''
-
-    @property
-    def is_link(self):
-        return (self.obj is not None)
-
-    @property
-    def is_list(self):
-        return self.obj and type(self.obj) == list
-
-    @property
-    def object_list(self):
-        if self.is_list:
-            return self.obj
-        else:
-            return []
-
-    @property
-    def link_id(self):
-        if not self.is_link:
-            return None
-
-        if issubclass(self.obj.__class__, MDR.aristotleComponent):
-            return self.obj.parentItem.id
-        else:
-            return self.obj.id
-
-    def __str__(self):
-        if self.is_link:
-            return self.obj.name
-        else:
-            return self.value or 'None'
-
-
-class ConceptVersionView(ConceptRenderMixin, TemplateView):
-
-    slug_redirect = False
-    version_arg = 'verid'
-    template_name = 'aristotle_mdr/concepts/managedContentVersion.html'
-    concept_fields = ['references', 'submitting_organisation', 'responsible_organisation',
-                      'origin', 'origin_URI', 'comments']
-    default_weak_map = {
-        'aristotle_mdr_slots.slot': 'concept',
-        'aristotle_mdr_identifiers.scopedidentifier': 'concept'
-    }
-
-    def check_item(self, item):
-        # Will 403 when user cant view the item
-        return user_can_view(self.request.user, item)
-
-    def get_item(self):
-        # Gets the current item
-        return self.item_version.object
-
-    def get_matching_object_from_revision(self, revision, current_version, target_ct=None):
-        # Finds another version in the same revision with same id
-        current_ct = current_version.content_type
-        version_filter = Q(revision=revision) &\
-            Q(object_id=current_version.object_id) &\
-            ~Q(content_type_id=current_ct.id)
-
-        # Other versions in the revision could have the same id
-        versions = reversion.models.Version.objects.filter(
-            version_filter
-        )
-
-        target_version = None
-        for sub_version in versions:
-            ct = sub_version.content_type
-            if target_ct is not None:
-                if sub_version.content_type == target_ct:
-                    target_version = sub_version
-                    break
-            else:
-                if issubclass(ct.model_class(), MDR._concept):
-                    # Find version that is a _concept subclass
-                    # Since the pk is the _concept_ptr this is fine
-                    target_version = sub_version
-                    break
-
-        return target_version
-
-    def get_version(self):
-        # Get the version of a concept and its matching subclass
-        try:
-            version = reversion.models.Version.objects.get(id=self.kwargs[self.version_arg])
-        except reversion.models.Version.DoesNotExist:
-            return False
-
-        self.revision = version.revision
-        concept_ct = ContentType.objects.get_for_model(MDR._concept)
-
-        # If we got a concept version id
-        if version.content_type == concept_ct:
-            self.concept_version = version
-            self.item_version = self.get_matching_object_from_revision(
-                self.revision,
-                version
-            )
-            if self.item_version is None:
-                return False
-        # If we got a concept subclass's version id
-        elif issubclass(version.content_type.model_class(), MDR._concept):
-            self.item_version = version
-            self.concept_version = self.get_matching_object_from_revision(
-                self.revision,
-                version,
-                concept_ct
-            )
-        else:
-            return False
-
-        self.concept_version_data = json.loads(self.concept_version.serialized_data)[0]
-        self.item_version_data = json.loads(self.item_version.serialized_data)[0]
-        self.item_model = self.item_version.content_type.model_class()
-
-        return True
-
-    def get_weak_versions(self, model):
-        # Get version data for weak entites (reverse relations to an
-        # aristotleComponent)
-
-        pk = self.item_version_data['pk']
-
-        # Find weak models create mapping of model labels to link fields
-        weak_map = self.default_weak_map
-        for field in model._meta.get_fields():
-            if field.is_relation and field.one_to_many and\
-                    issubclass(field.related_model, MDR.aristotleComponent):
-                weak_map[field.related_model._meta.label_lower] = field.field.name
-
-        if len(weak_map) == 0:
-            return []
-
-        # Get any models found before in the same revision to dict
-        weak_items = self.get_related_versions(pk, weak_map)
-
-        # Process into template friendly version
-        template_weak_models = []
-        for label, item_dict in weak_items.items():
-            model = apps.get_model(label)
-
-            template_weak_models.append({
-                'model': pretify_camel_case(model.__name__),
-                'headers': item_dict['headers'],
-                'items': item_dict['items']
-            })
-
-        return template_weak_models
-
-    def process_dict(self, fields, model):
-        # Process fields dict, updating field names and links
-
-        # Create replacement mapping fields to models
-        replacements = {}
-        for field in model._meta.get_fields():
-            if field.is_relation and (field.many_to_one or field.many_to_many) and\
-                    (issubclass(field.related_model, MDR._concept) or
-                        issubclass(field.related_model, MDR.aristotleComponent)):
-                replacements[field.name] = field.related_model
-
-        # Create new mapping with user friendly keys and replaced models
-        updated_fields = {}
-        for key, value in fields.items():
-            field = model._meta.get_field(key)
-            header = field.verbose_name.title()
-
-            replaced = False
-            if key in replacements:
-
-                sub_model = replacements[key]
-
-                if type(value) == int and field.many_to_one:
-                    updated_fields[header] = self.lookup_object([value], sub_model, field)
-                    replaced = True
-                elif type(value) == list and field.many_to_many and value:
-                    updated_fields[header] = self.lookup_object(value, sub_model, field)
-                    replaced = True
-
-            if not replaced:
-                updated_fields[header] = VersionField(
-                    value=value,
-                    help_text=field.help_text
-                )
-
-        return updated_fields
-
-    def lookup_object(self, pk_list, sub_model, field):
-
-        if issubclass(sub_model, MDR._concept):
-            label = MDR._concept._meta.label_lower
-        else:
-            label = sub_model._meta.label_lower
-
-        self.obj_ids[label] += pk_list
-
-        ver_field = VersionField(
-            help_text=field.help_text,
-            obj=pk_list,
-            reference_label=label
-        )
-
-        return ver_field
-
-    def lookup_object_refs(self):
-        for label, id_list in self.obj_ids.items():
-            model = apps.get_model(label)
-            object_qs = model.objects.filter(id__in=id_list)
-            object_map = {}
-            for obj in object_qs:
-                object_map[obj.id] = obj
-
-            self.fetched_objects[label] = object_map
-
-    def replace_object_refs(self, field_dict):
-        for key in field_dict.keys():
-            field_dict[key].dereference(self.fetched_objects)
-
-        return field_dict
-
-    def get_related_versions(self, pk, mapping):
-        # mapping should be a mapping of model labels to fields on item
-
-        related = {}
-        # Add any models found before in the same revision to dict
-        for version in self.revision.version_set.all():
-            data = json.loads(version.serialized_data)[0]
-            if data['model'] in mapping:
-
-                if data['model'] not in related:
-                    related[data['model']] = {
-                        'items': []
-                    }
-
-                # There is a version in the revision that is of the correct
-                # type. Need to check wether it links to the correct item
-                related_model = apps.get_model(data['model'])
-
-                # Find the field that links the weak model back to our model
-                link_field = mapping[data['model']]
-
-                if link_field and link_field in data['fields']:
-                    # If it links back to the correct pk
-                    if data['fields'][link_field] == pk:
-                        # Add to weak models
-                        del data['fields'][link_field]
-                        final_fields = self.process_dict(data['fields'], related_model)
-
-                        if 'headers' not in related[data['model']]:
-                            headers = []
-                            for header, item in final_fields.items():
-                                headers.append({
-                                    'text': header,
-                                    'help_text': item.help_text
-                                })
-                            related[data['model']]['headers'] = headers
-
-                        related[data['model']]['items'].append(final_fields)
-
-        return related
-
-    def dispatch(self, request, *args, **kwargs):
-        # Dict mapping models to a list of ids
-        self.obj_ids = defaultdict(list)
-        # Dict mapping model -> id -> object
-        self.fetched_objects = {}
-
-        exists = self.get_version()
-        if not exists:
-            return HttpResponseNotFound()
-
-        return super().dispatch(request, *args, **kwargs)
-
-    def get_version_context_data(self):
-        # Get the context data for this complete version
-
-        version_dict = self.concept_version_data['fields']
-        # Keys under item_data are used as headings
-        version_dict['item_data'] = {'Names & References': {}}
-
-        # Replace workgroup reference with wg object
-        if version_dict['workgroup']:
-            try:
-                workgroup = MDR.Workgroup.objects.get(pk=version_dict['workgroup'])
-            except MDR.Workgroup.DoesNotExist:
-                workgroup = None
-
-            version_dict['workgroup'] = workgroup
-
-        # Add concept fields as "Names & References"
-        for field in self.concept_fields:
-            if field in self.concept_version_data['fields']:
-                fieldobj = MDR._concept._meta.get_field(field)
-                is_html = (issubclass(type(fieldobj), RichTextField))
-                field = VersionField(
-                    value=self.concept_version_data['fields'][field],
-                    help_text=fieldobj.help_text,
-                    html=is_html
-                )
-
-                # Keys under item_data are used as headings
-                version_dict['item_data']['Names & References'][fieldobj.verbose_name.title()] = field
-
-        # Add some extra data the temlate expects from a regular item object
-        version_dict['meta'] = {
-            'app_label': self.item_version.content_type.app_label,
-            'model_name': self.item_version.content_type.model
-        }
-        version_dict['id'] = self.item_version_data['pk']
-        version_dict['pk'] = self.item_version_data['pk']
-        version_dict['get_verbose_name'] = self.item_version.content_type.name.title()
-        version_dict['created'] = parse_datetime(self.concept_version_data['fields']['created'])
-
-        # Add weak entities and components
-        weak = self.get_weak_versions(self.item_model)
-        components = self.process_dict(self.item_version_data['fields'], self.item_model)
-        self.lookup_object_refs()
-        version_dict['item_data']['Components'] = self.replace_object_refs(components)
-
-        for i in range(len(weak)):
-            for j in range(len(weak[i]['items'])):
-                weak[i]['items'][j] = self.replace_object_refs(weak[i]['items'][j])
-
-        version_dict['weak'] = weak
-
-        return version_dict
-
-    def get_context_data(self, *args, **kwargs):
-        context = super(ConceptRenderMixin, self).get_context_data(*args, **kwargs)
-        context['hide_item_actions'] = True
-        context['hide_item_supersedes'] = True
-        context['hide_item_help'] = True
-        context['hide_item_related'] = True
-        context['item'] = self.get_version_context_data()
-        context['current_item'] = self.item
-        context['revision'] = self.revision
-        context['item_is_version'] = True
-        return context
-
-    def get_template_names(self):
-        return [self.template_name]
 
 
 def registrationHistory(request, iid):
